Wörterbuch is now functioning, mostly
I have made it where as one types in the combobox, that words are
suggested, much like on google search. However it doesn't seem to
register the first time a letter is typed after starting up. But after
the second letter is typed, it works fine. I should make the search
results given by the dataProvider class only 4, so that, when 5k words
that begin with A, don't all have to be filled into the List<>.